Ignition server: Use https health check if possible
Currently, the tcp health checking spams the ignition server log with
error. This change adds a simple http healthz endpoint and uses that for
healthchecking if the docker image has a specific label that is also
added. Featuregating this is needed, because the new healthcheck will
never succeed on older ignition servers that do not have this new
handler.